+++ Sources/DP3TSDK/Networking/ExposeeServiceClient.swift
@@ -185,9 +185,16 @@ class ExposeeServiceClient: ExposeeServiceClientProtocol {
         request.addValue("application/json", forHTTPHeaderField: "Content-Type")
         request.addValue(String(payload.count), forHTTPHeaderField: "Content-Length")
         request.addValue(userAgent, forHTTPHeaderField: "User-Agent")
-        if case let ExposeeAuthMethod.HTTPAuthorizationBearer(token: token) = authentication {
+
+        switch authentication {
+        case .none:
+            break
+        case let .HTTPAuthorizationHeader(header: header, value: value):
+            request.addValue(value, forHTTPHeaderField: header)
+        case let .HTTPAuthorizationBearer(token: token):
             request.addValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
         }
+
         request.httpBody = payload
 
         log.log("addExposeeList URL: %{public}@", request.url?.absoluteString ?? "nil")